refactor: move application class to its own file
