+++ src/Application.h
@@ -1,0 +1,78 @@
+/******************************************************************************
+ * @file   Application.h
+ * @author Tim Brewis (@t-bre, tab1g19@soton.ac.uk)
+ * @brief  Main JUCE application
+ *****************************************************************************/
+
+#pragma once
+
+#include <JuceHeader.h>
+#include <limits.h>
+
+#include "MainComponent.h"
+
+/**
+ * @brief Main JUCE application
+ */
+class Application : public juce::JUCEApplication
+{
+public:
+    
+    Application() = default;
+
+    void initialise(const juce::String& commandLine) override;
+    void shutdown() override;
+    void systemRequestedQuit() override;
+    void anotherInstanceStarted(const juce::String& commandLine) override;
+
+    const juce::String getApplicationName() override;
+    const juce::String getApplicationVersion() override;
+    bool moreThanOneInstanceAllowed() override;
+
+    //==============================================================================
+    /*
+        This class implements the desktop window that contains an instance of
+        our MainComponent class.
+    */
+    class MainWindow : public juce::DocumentWindow
+    {
+      public:
+        MainWindow(juce::String name)
+            : DocumentWindow(
+                name,
+                juce::Desktop::getInstance().getDefaultLookAndFeel().findColour(
+                    juce::ResizableWindow::backgroundColourId),
+                DocumentWindow::allButtons)
+        {
+            setUsingNativeTitleBar(true);
+            setContentOwned(new MainComponent(), true);
+            setResizeLimits(minWidth, minHeight, INT_MAX, INT_MAX);
+
+#if JUCE_IOS || JUCE_ANDROID
+            setFullScreen(true);
+#else
+            setResizable(true, true);
+            centreWithSize(getWidth(), getHeight());
+#endif
+
+            setVisible(true);
+        }
+
+        void closeButtonPressed() override
+        {
+            JUCEApplication::getInstance()->systemRequestedQuit();
+        }
+
+      private:
+        // constants
+        static const int minWidth = 500;
+        static const int minHeight = 350;
+
+        JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(MainWindow)
+    };
+
+private:
+
+    std::unique_ptr<MainWindow> mainWindow;
+
+};
