+++ src/Main.cpp
@@ -1,114 +1,11 @@
 /******************************************************************************
  * @file   Main.cpp
- * @author Tim Brewis (tab1g19@soton.ac.uk)
- * @date   2022-02-09
- * @brief  JUCE application
+ * @author Tim Brewis (@t-bre tab1g19@soton.ac.uk)
+ * @brief  Generates the main() routine that launches the app
  *****************************************************************************/
 
-#include "MainComponent.h"
 #include <JuceHeader.h>
-#include <limits.h>
 
-//==============================================================================
-class VCUDriverProfileApplication : public juce::JUCEApplication
-{
-  public:
-    //==============================================================================
-    VCUDriverProfileApplication()
-    {
-    }
+#include "Application.h"
 
-    const juce::String getApplicationName() override
-    {
-        return ProjectInfo::projectName;
-    }
-    const juce::String getApplicationVersion() override
-    {
-        return ProjectInfo::versionString;
-    }
-    bool moreThanOneInstanceAllowed() override
-    {
-        return false;
-    }
-
-    //==============================================================================
-    void initialise(const juce::String& commandLine) override
-    {
-        // This method is where you should put your application's initialisation
-        // code..
-
-        mainWindow.reset(new MainWindow("VCU Driver Profile"));
-    }
-
-    void shutdown() override
-    {
-        // Add your application's shutdown code here..
-
-        mainWindow = nullptr; // (deletes our window)
-    }
-
-    //==============================================================================
-    void systemRequestedQuit() override
-    {
-        // This is called when the app is being asked to quit: you can ignore
-        // this request and let the app carry on running, or call quit() to
-        // allow the app to close.
-        quit();
-    }
-
-    void anotherInstanceStarted(const juce::String& commandLine) override
-    {
-        // When another instance of the app is launched while this one is
-        // running, this method is invoked, and the commandLine parameter tells
-        // you what the other instance's command-line arguments were.
-    }
-
-    //==============================================================================
-    /*
-        This class implements the desktop window that contains an instance of
-        our MainComponent class.
-    */
-    class MainWindow : public juce::DocumentWindow
-    {
-      public:
-        MainWindow(juce::String name)
-            : DocumentWindow(
-                name,
-                juce::Desktop::getInstance().getDefaultLookAndFeel().findColour(
-                    juce::ResizableWindow::backgroundColourId),
-                DocumentWindow::allButtons)
-        {
-            setUsingNativeTitleBar(true);
-            setContentOwned(new MainComponent(), true);
-            setResizeLimits(minWidth, minHeight, INT_MAX, INT_MAX);
-
-#if JUCE_IOS || JUCE_ANDROID
-            setFullScreen(true);
-#else
-            setResizable(true, true);
-            centreWithSize(getWidth(), getHeight());
-#endif
-
-            setVisible(true);
-        }
-
-        void closeButtonPressed() override
-        {
-            JUCEApplication::getInstance()->systemRequestedQuit();
-        }
-
-      private:
-        // constants
-        static const int minWidth = 500;
-        static const int minHeight = 350;
-
-        JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(MainWindow)
-    };
-
-  private:
-    std::unique_ptr<MainWindow> mainWindow;
-};
-
-//==============================================================================
-// This macro generates the main() routine that launches the app.
-START_JUCE_APPLICATION(VCUDriverProfileApplication)
+START_JUCE_APPLICATION(Application)
